requirements/common.txt
pandas==0.24.2
Flask==1.0.2
Flask_Cors==3.0.7
pytest==4.5.0
tensorflow==1.13.1
Flask_SQLAlchemy==2.3.2
numpy==1.16.4
scipy==1.2.1
SQLAlchemy==1.3.3
typing==3.6.4
Flask_Migrate==2.4.0
Rx==3.0.0b2
python-dotenv==0.10.3
scikit_learn==0.21.2
